+++ dotnet/Services/ProductReviewService.cs
@@ -115,9 +115,14 @@
             return review;
         }
 
-        public async Task<decimal> GetAverageRatingByProductId(string productId)
+        public async Task<AverageCount> GetAverageRatingByProductId(string productId)
         {
             decimal averageRating = 0m;
+            int stars1 = 0;
+            int stars2 = 0;
+            int stars3 = 0;
+            int stars4 = 0;
+            int stars5 = 0;
 
             string searchQuery = $"productId={productId}";
             AppSettings settings = await GetAppSettings();
@@ -133,12 +138,33 @@
                 int numberOfReviews = reviews.Count;
                 if (numberOfReviews > 0)
                 {
-                    decimal totalRating = reviews.Sum(r => r.Rating ?? 0);
+                    decimal totalRating = 0;
+                    foreach(Review review in reviews)
+                    {
+                        if (review.Rating != null)
+                        {
+                            totalRating = totalRating + review.Rating ?? 0;
+                            if (review.Rating == 5) stars5++;
+                            else if (review.Rating == 4) stars4++;
+                            else if (review.Rating == 3) stars3++;
+                            else if (review.Rating == 2) stars2++;
+                            else stars1++;
+                        }
+                    }
                     averageRating = totalRating / numberOfReviews;
                 }
             }
 
-            return decimal.Round(averageRating, 2, MidpointRounding.AwayFromZero);
+            AverageCount avergae = new AverageCount
+            {
+                Average = decimal.Round(averageRating, 2, MidpointRounding.AwayFromZero),
+                StarsFive = stars5,
+                StarsFour = stars4,
+                StarsThree = stars3,
+                StarsTwo = stars2,
+                StarsOne = stars1
+            };
+            return avergae;
         }
 
         public async Task<Review> GetReview(string Id)